fix(core,min/max): initial value was incorrect
